refactor: review 전체 조회 dto로 한번 더 감싸기
