+++ backend/src/main/java/zipgo/petfood/presentation/PetFoodController.java
@@ -1,19 +1,9 @@
 package zipgo.petfood.presentation;
 
-import static java.net.URLDecoder.decode;
-import static java.util.Collections.EMPTY_LIST;
-
 import io.jsonwebtoken.lang.Strings;
-import java.io.UnsupportedEncodingException;
-import java.util.Arrays;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import zipgo.brand.domain.Brand;
 import zipgo.petfood.application.PetFoodQueryService;
 import zipgo.petfood.domain.PetFood;
@@ -21,6 +11,13 @@ import zipgo.petfood.presentation.dto.FilterMetadataResponse;
 import zipgo.petfood.presentation.dto.FilterResponse;
 import zipgo.petfood.presentation.dto.GetPetFoodResponse;
 import zipgo.petfood.presentation.dto.GetPetFoodsResponse;
+
+import java.io.UnsupportedEncodingException;
+import java.util.Arrays;
+import java.util.List;
+
+import static java.net.URLDecoder.decode;
+import static java.util.Collections.EMPTY_LIST;
 
 @RestController
 @RequiredArgsConstructor
