+++ backend/src/main/java/zipgo/review/dto/response/GetReviewsResponse.java
@@ -1,38 +1,13 @@
 package zipgo.review.dto.response;
 
-import static java.time.format.DateTimeFormatter.ofPattern;
-
 import java.util.List;
-import zipgo.review.domain.Review;
 
 public record GetReviewsResponse(
-        Long id,
-        String reviewerName,
-        int rating,
-        String date,
-        String comment,
-        String tastePreference,
-        String stoolCondition,
-        List<String> adverseReactions
+        List<GetReviewResponse> reviews
 ) {
 
-    public static GetReviewsResponse from(Review review) {
-        return new GetReviewsResponse(
-                review.getId(),
-                review.getMember().getName(),
-                review.getRating(),
-                review.getCreatedAt().format(ofPattern("yyyy-MM-dd")),
-                review.getComment(),
-                review.getStoolCondition().getDescription(),
-                review.getTastePreference().getDescription(),
-                getAdverseReactions(review)
-        );
-    }
-
-    private static List<String> getAdverseReactions(Review review) {
-        return review.getAdverseReactions().stream()
-                .map(adverseReaction -> adverseReaction.getAdverseReactionType().getDescription())
-                .toList();
+    public static GetReviewsResponse from(List<GetReviewResponse> reviews) {
+        return new GetReviewsResponse(reviews);
     }
 
 }
